fix problem of delete comment table

+++ trunk/jsboard-2/admin/act.php
@@ -41,9 +41,11 @@ if($mode=='db_del') {
   mysql_query($table_del,$connect);
   sql_error(mysql_errno(),mysql_error());
 
-  $comm_del = "drop table {$table_name}_comm";
-  mysql_query($comm_del,$connect);
-  sql_error(mysql_errno(),mysql_error());
+  if(!preg_match("/_comm$/",$table_name)) {
+    $comm_del = "drop table {$table_name}_comm";
+    mysql_query($comm_del,$connect);
+    sql_error(mysql_errno(),mysql_error());
+  }
 
   # 게시판 계정에서 사용되는 file 삭제
   exec("$exec[rm] ../data/$table_name");
